Mashup of various updates related to existing display
New behavior allows specifying display-viewport height and width. This is account for various apps. and how they handle representing a site on the Chromecast.
Some display as-expected, while others do weird vertical-scaling (i.e. Web2Cast).
New behavior also allows specifying an audio source on the existing display.
And, create a development config. file which is intended to NOT be committed to repo.